requirements.txt
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.5.2
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
azure-mgmt-compute==34.1.0
azure-mgmt-core==1.5.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-monitor==6.0.2
azure-mgmt-resource==23.3.0
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.5
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.9.1
msal==1.35.1
msal-extensions==1.3.0
pycparser==2.23
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
typing-extensions==4.13.2
urllib3==2.2.3
